+++ src/llmcompressor/modifiers/transform/spinquant/__init__.py
@@ -1,1 +1,3 @@
+# flake8: noqa
+
 from .base import *
